+++ GHEtool/VariableClasses/PipeData/__init__.py
@@ -1,4 +1,4 @@
 from ._PipeData import _PipeData
 from .SingleUPipe import SingleUPipe
 from .DoubleUPipe import DoubleUPipe
-from .NUPipe import NUPipe
+from .MultipleUTube import MultipleUTube