+++ CHANGELOG.md
@@ -9,6 +9,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Extra warning message if one wants to load a GHEtool file that was created with a newer version.
 - Borehole thermal resistance is now visible at the borehole thermal resistance page (issue #51).
 - New class of GroundData: GroundTemperatureGradient added (issue #145).
+- Load classes (issue #45).
+- Pipe classes (single, double, coaxial, Multiple U Tube) (issue #45).
 
 ## Changed
 - GUI was moved to a seperate project: ScenarioGUI.